# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: CRM module for Uralintern
  description: API documentation for Uralintern CRM Module
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: kayakto server url
security:
  - BearerAuth: []
...
paths:
...
  /events-curators/{eventId}/send/{curatorId}:
    put:
      tags:
        - event-curator-controller
      operationId: sendCuratorToEvent
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: curatorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events-curators/{eventId}/reject/{curatorId}:
    put:
      tags:
        - event-curator-controller
      operationId: rejectCuratorRequest
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: curatorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events-curators/{eventId}/accept/{curatorId}:
    put:
      tags:
        - event-curator-controller
      operationId: acceptCuratorRequest
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: curatorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /events-curators/{eventId}/waiting-curators:
    get:
      tags:
        - event-curator-controller
      operationId: getSentCuratorsInfo
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events-curators/{eventId}/curators:
    get:
      tags:
        - event-curator-controller
      operationId: getCuratorsInfo
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events-curators/{eventId}/curator-status/{curatorId}:
    get:
      tags:
        - event-curator-controller
      operationId: getCuratorStatus
      parameters:
        - name: eventId
          in: query
          required: true
          schema:
            type: integer
            format: int64
            example: 1
        - name: curatorId
          in: query
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events-curators/{eventId}/accepted-curators:
    get:
      tags:
        - event-curator-controller
      operationId: getAcceptedCuratorsInfo
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /events-curators/{eventId}/delete/{curatorId}:
    delete:
      tags:
        - event-curator-controller
      operationId: deleteCuratorFromEvent
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: curatorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
components:
  responses:
      OkResponse:
        description: OK, запрос выполнен успешно
        content:
          '*/*':
            schema:
              type: boolean
      BadRequestResponse:
        description: Bad Request, запрос неверный со стороны клиента
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ErrorResponse'
      NotFoundResponse:
        description: Not Found, такого элемента не существует
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ErrorResponse'
      InternalServerErrorResponse:
        description: Internal Server Error, запрос вызвал ошибку на сервере
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "User with email email123@gmail.com not found"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT